page name change to FlightPage
