+++ front/search-flights/src/pages/FlightsPage/FP_sxStyles.tsx
@@ -1,10 +1,5 @@
 export const boxesContainerSxStyle = { display: "flex" };
 
-export const containerTop = {
-  display: "flex",
-  marginTop: 2,
-  gap: 4,
-};
 export const leftBoxSxStyle = {
   width: "35%",
   paddingTop: 2,
